+++ frontend/lib/widgets/common/main_app_bar.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
-import '../theme/colors.dart';
-import '../theme/text_styles.dart';
+import '../../theme/colors.dart';
+import '../../theme/text_styles.dart';
 
 class MainAppBar extends StatelessWidget implements PreferredSizeWidget {
   final String? title;
