+++ frontend/lib/widgets/common/mode_switcher.dart
