+++ frontend/lib/widgets/common/search_filter_chips.dart
