+++ frontend/lib/widgets/dashboard/widgets.dart
@@ -1,0 +1,2 @@
+export 'user_info_card.dart';
+export 'dashboard_tabs.dart';
