+++ frontend/lib/widgets/discover/project_card.dart
