+++ frontend/lib/widgets/discover/search_bar.dart
