+++ frontend/lib/widgets/discover/talent_card.dart
