+++ frontend/lib/widgets/discover/widgets.dart
@@ -1,0 +1,4 @@
+export 'project_card.dart';
+export 'talent_card.dart';
+export 'search_bar.dart';
+export 'sort_filter_button.dart';
