+++ frontend/lib/widgets/login/tab_selector.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
-import '../theme/colors.dart';
-import '../theme/dimensions.dart';
+import '../../theme/colors.dart';
+import '../../theme/dimensions.dart';
 
 class TabSelector extends StatelessWidget {
   final TabController controller;
