--- screens/pages/discover_page.dart
@@ -1,1 +1,0 @@
-
