--- screens/pages/settings_page.dart
@@ -1,1 +1,0 @@
-
