--- screens/pages/views/project_view.dart
@@ -1,1 +1,0 @@
-
