--- screens/pages/views/talent_view.dart
@@ -1,1 +1,0 @@
-
