--- theme/text_styles.dart
@@ -1,10 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:frontend/theme/colors.dart';
-
-class AppTextStyles {
-  static const TextStyle cardLabelStyle = TextStyle(
-    fontWeight: FontWeight.bold,
-    fontSize: 14,
-    color: AppColors.textPrimary,
-  );
-}
