--- widgets/talent_card.dart
@@ -1,1 +1,0 @@
-
